+++ pandas/tests/frame/indexing/test_setitem.py
@@ -728,36 +728,6 @@ class TestDataFrameSetItem:
         df.isetitem(0, DataFrame({"a": [10, 11]}, index=[1, 2]))
         tm.assert_frame_equal(df, expected)
 
-    def test_isetitem_ea_df(self):
-        # GH#49922
-        df = DataFrame([[1, 2, 3], [4, 5, 6]])
-        rhs = DataFrame([[11, 12], [13, 14]], dtype="Int64")
-
-        df.isetitem([0, 1], rhs)
-        expected = DataFrame(
-            {
-                0: Series([11, 13], dtype="Int64"),
-                1: Series([12, 14], dtype="Int64"),
-                2: [3, 6],
-            }
-        )
-        tm.assert_frame_equal(df, expected)
-
-    def test_isetitem_ea_df_scalar_indexer(self):
-        # GH#49922
-        df = DataFrame([[1, 2, 3], [4, 5, 6]])
-        rhs = DataFrame([[11], [13]], dtype="Int64")
-
-        df.isetitem(2, rhs)
-        expected = DataFrame(
-            {
-                0: [1, 4],
-                1: [2, 5],
-                2: Series([11, 13], dtype="Int64"),
-            }
-        )
-        tm.assert_frame_equal(df, expected)
-
     def test_setitem_frame_overwrite_with_ea_dtype(self, any_numeric_ea_dtype):
         # GH#46896
         df = DataFrame(columns=["a", "b"], data=[[1, 2], [3, 4]])
